Implemented a streak blur shader in special made by a person smarter than me :P

+++ KailashEngine/Render/Shader/glsl/special/special_BlurGuass.frag
@@ -3,10 +3,10 @@
 in vec2 v_TexCoord;
 
 
-uniform sampler2D sampler0;				// Source Texture
+uniform sampler2D sampler0;		// Source Texture
 
 
-uniform float blur_amount;
+uniform int blur_amount;
 uniform vec2 texture_size;
